+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # @destinygg/libstiny
 
+## 1.0.0
+
+### Major Changes
+
+- Added Tabs
+- Added Choicebox
+- Refactored Alert into Notification. Added Alert and Toast variants
+- Added `icon-only` Button and NavbarItem variants
+- Added `card__description`
+
 ## 0.0.8-alpha
 
 ### Patch Changes